Success (with features): Upload to Django was successful on VM
# There are features of the script: when preparing TXT files for uploading to Django, adjustments are made to the original files:
1. Removed "lbs" next to the weight of the fruit
2. The name of the photo is added to the end of the file

In files 1 and 10, after the text of the data in the TXT file, there is an Enter, in other files, it is not.
The script works correctly with files 2-9 and does not work with files 1 and 10.

!!! To successfully complete part 3 of the lab, manually run the following commands

nano ~/supplier-data/descriptions/010.txt
nano ~/supplier-data/descriptions/001.txt

and manually erase the extra ("not typical") Enter at the end of the TXT file.

what you need to run the script run.py
!!! before running run.py, you need to run the script run_first.py 1 time